Implemented PathFinder test cases
